requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
arrow==1.3.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colour==0.1.5
cpe==1.3.1
cryptography==44.0.0
cybox==2.1.0.21
deepdiff==8.4.2
distro==1.9.0
drawsvg==2.4.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.3.1
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llamaapi==0.1.36
loguru==0.7.3
lxml==5.3.2
maec==4.1.0.17
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mitreattack-python==3.0.8
mixbox==1.0.5
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.4.2
numpy==2.2.3
openai==1.60.1
openpyxl==3.1.5
ordered-set==4.1.0
orderly-set==5.4.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.7
pluralizer==1.2.0
pooch==1.8.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.2.0
stixmarx==1.0.8
sympy==1.13.3
tabulate==0.9.0
taxii2-client==2.3.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
weakrefmethod==1.0.3
webcolors==24.11.1
websockets==15.0.1
Werkzeug==3.1.3
XlsxWriter==3.2.2
yarl==1.18.3
